Modified the Checkpoint record to use fixed length fields in the record in order to better support unversioned updates.  Visited more of the tuple serializers to make sure that they are also versioned.
git-svn-id: svn+ssh://svn.code.sf.net/p/bigdata/code/trunk@3318 8f7bc0f5-282e-0410-95e3-8d296e9bb460

+++ bigdata/src/java/com/bigdata/btree/DefaultTupleSerializer.java
@@ -314,17 +314,17 @@ public class DefaultTupleSerializer<K extends Object, V extends Object>
      * Note: Explicit versioning for the {@link DefaultTupleSerializer} was
      * introduced with inlining of datatype literals for the RDF database.
      */
-    private final static transient int VERSION0 = 0;
+    private final static transient byte VERSION0 = 0;
 
     /**
      * The current version.
      */
-    private final static transient int VERSION = VERSION0;
+    private final static transient byte VERSION = VERSION0;
 
     public void readExternal(final ObjectInput in) throws IOException,
             ClassNotFoundException {
 
-        final short version = in.readShort();
+        final byte version = in.readByte();
         switch (version) {
         case VERSION0:
             delegateKeyBuilderFactory = (IKeyBuilderFactory) in.readObject();
@@ -342,7 +342,7 @@ public class DefaultTupleSerializer<K extends Object, V extends Object>
 
     public void writeExternal(final ObjectOutput out) throws IOException {
 
-        out.writeShort(VERSION);
+        out.writeByte(VERSION);
         
         out.writeObject(delegateKeyBuilderFactory);
